separated off menu view
